refactor: Flatten plugins/interflow structure
Moves inter-flow policy implementations directly into the
`plugins/interflow` package, removing the unnecessary nested
directories.

This simplifies the import paths and project structure.

This is a no-op refactoring.
